'catts' and 'tsana' support 64-bit linux now, 'catip' has 'struct ip_mreq' problem

+++ include/libts/if.h
@@ -16,7 +16,7 @@ extern "C" {
 int b2t(char *DST, const uint8_t *PTR, int len);
 int next_tag(char **tag, char **text);
 int next_nbyte_hex(uint8_t *byte, char **text, int max);
-int next_nuint_hex(int64_t *sint, char **text, int max);
+int next_nuint_hex(long long int *sint, char **text, int max);
 
 #ifdef __cplusplus
 }
